minor refactoring and test
