+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_SERVER=http://localhost:5000
