+++ src/components/styleElements/icons/Check.js
@@ -5,7 +5,7 @@ import './Icon.css';
 
 export function Check({ width, height }) {
   return (
-    <svg
+    <svg data-testid="check"
       width={width ? width : 18}
       height={height ? height : 13}
       viewBox="0 0 18 13"
