+++ src/components/styleElements/icons/X.js
@@ -3,7 +3,7 @@ import './Icon.css';
 
 export const X = ({ size }) => {
   return (
-    <svg
+    <svg data-testid="x-icon"
       width={size ? size : null}
       height={size ? size : null}
       className={size ? null : 'iconSvg'}
